# Create goroutines adaptively during metrics gathering