Attempt to fix #369. This needs to be tested with one of those scenarios as in https://github.com/prometheus/statsd_exporter/issues/76 .

@grobie @brian-brazil