@matthiasr 2018-01-28
prometheus/statsd_exporter#119 has benchmarks and profiling that you can use to test the effects of this.