@brian-brazil 2018-01-29
The code seems okay from a first lookthrough. I'd worry that this approach might cause unpredictable goroutine/memory usage when a varying number of collectors are slow. Maybe limit the number of goroutines to the number of cores or something to bound it?