@beorn7 2018-01-29
@matthiasr I'll check out the statsd_exporter benchmarks as a test case for this.

@brian-brazil the number of goroutine is effectively limited here to the number of collectors. Limiting it to the number of cores would not make sense. It's the whole point of the go execution model to have more goroutines than cores. E.g. if a goroutine is waiting for I/O, it gets effectively detached from its current OS thread, and the latter will run another goroutine.