@beorn7 2018-01-29
Benchmark result from https://github.com/prometheus/statsd_exporter/pull/119 :

BEFORE:
```
BenchmarkGatherGauge/m_100000_l_1-4                    2         634044750 ns/op        61646164 B/op    1157861 allocs/op
BenchmarkGatherGauge/m_10000_l_10-4                    3         342434786 ns/op        26507605 B/op     514057 allocs/op
BenchmarkGatherGauge/m_10_l_10000-4                    5         296876316 ns/op        23959355 B/op     404184 allocs/op
BenchmarkGatherGauge/m_1_l_100000-4                    3         496134591 ns/op        24761392 B/op     404040 allocs/op
BenchmarkGatherCounter/m_100000_l_1-4                  2         550082110 ns/op        61584880 B/op    1157420 allocs/op
BenchmarkGatherCounter/m_10000_l_10-4                  3         354095715 ns/op        26505234 B/op     514031 allocs/op
BenchmarkGatherCounter/m_10_l_10000-4                  5         294096041 ns/op        23966792 B/op     404187 allocs/op
BenchmarkGatherCounter/m_1_l_100000-4                  3         481249290 ns/op        24753322 B/op     404016 allocs/op
BenchmarkGatherSummary/m_100000_l_1-4                  1        1502250708 ns/op        105471992 B/op   2606115 allocs/op
BenchmarkGatherSummary/m_10000_l_10-4                  1        1237115864 ns/op        60737560 B/op    1833445 allocs/op
BenchmarkGatherSummary/m_10_l_10000-4                  2         603992947 ns/op        53568084 B/op    1604204 allocs/op
BenchmarkGatherSummary/m_1_l_100000-4                  1        1067372223 ns/op        57537016 B/op    1703965 allocs/op
BenchmarkGatherHistogram/m_100000_l_1-4                1        1748068925 ns/op        156688928 B/op   4805885 allocs/op
BenchmarkGatherHistogram/m_10000_l_10-4                2         585414979 ns/op        110880208 B/op   4018966 allocs/op
BenchmarkGatherHistogram/m_10_l_10000-4                2         538948882 ns/op        107956244 B/op   3904172 allocs/op
BenchmarkGatherHistogram/m_1_l_100000-4                2         704194725 ns/op        108760040 B/op   3904024 allocs/op
```

AFTER (i.e. with this PR in):
```
BenchmarkGatherGauge/m_100000_l_1-4                    3         440166007 ns/op        52157970 B/op    1009594 allocs/op
BenchmarkGatherGauge/m_10000_l_10-4                   10         172004623 ns/op        25938732 B/op     504239 allocs/op
BenchmarkGatherGauge/m_10_l_10000-4                    5         263151519 ns/op        23970416 B/op     404233 allocs/op
BenchmarkGatherGauge/m_1_l_100000-4                    3         482181690 ns/op        24759512 B/op     404018 allocs/op
BenchmarkGatherCounter/m_100000_l_1-4                  3         437489358 ns/op        52168122 B/op    1009618 allocs/op
BenchmarkGatherCounter/m_10000_l_10-4                 10         182594454 ns/op        25941801 B/op     504233 allocs/op
BenchmarkGatherCounter/m_10_l_10000-4                  3         349815703 ns/op        23961765 B/op     404202 allocs/op
BenchmarkGatherCounter/m_1_l_100000-4                  3         483724620 ns/op        24746320 B/op     404000 allocs/op
BenchmarkGatherSummary/m_100000_l_1-4                  1        1298970300 ns/op        84972176 B/op    2309705 allocs/op
BenchmarkGatherSummary/m_10000_l_10-4                  5         286344748 ns/op        55537083 B/op    1704213 allocs/op
BenchmarkGatherSummary/m_10_l_10000-4                  3         446114870 ns/op        53563317 B/op    1604220 allocs/op
BenchmarkGatherSummary/m_1_l_100000-4                  2         883968198 ns/op        54331852 B/op    1603902 allocs/op
BenchmarkGatherHistogram/m_100000_l_1-4                1        1009265108 ns/op        136181704 B/op   4509751 allocs/op
BenchmarkGatherHistogram/m_10000_l_10-4                3         466298888 ns/op        109941794 B/op   4004212 allocs/op
BenchmarkGatherHistogram/m_10_l_10000-4                2         558274380 ns/op        107960956 B/op   3904228 allocs/op
BenchmarkGatherHistogram/m_1_l_100000-4                2         891237764 ns/op        108757552 B/op   3903994 allocs/op
```

The cases with many individual collectors to gather from (`m_100000`) are treated more efficiently but not dramatically so. @matthiasr told me it gets interesting with a million and more, but that doesn't work with summaries (OOMs). Will run that for a Counter only. Also, looking at the memory profile is also interesting.