@beorn7 2018-01-29
Counter with a million:

BEFORE:
```
BenchmarkGatherCounter/m_1000000_l_1-4                 1        11464764011 ns/op       1236364384 B/op 14072161 allocs/op
BenchmarkGatherCounter/m_1_l_1000000-4                 1        7028819377 ns/op        263177648 B/op   4038419 allocs/op
```

AFTER:
```
BenchmarkGatherCounter/m_1000000_l_1-4                 1        4894638356 ns/op        599587720 B/op  10076272 allocs/op
BenchmarkGatherCounter/m_1_l_1000000-4                 1        7299555268 ns/op        263168624 B/op   4038438 allocs/op
```

Interestingly, gathering from 1M collectors is now even faster than gathering from one collector with a million metrics.