@beorn7 2018-01-29
Profiles to look at:

[m_1_l_1000000.cpu.BEFORE.pb.gz](https://github.com/prometheus/statsd_exporter/files/1674498/m_1_l_1000000.cpu.BEFORE.pb.gz)
[m_1_l_1000000.cpu.pb.gz](https://github.com/prometheus/statsd_exporter/files/1674499/m_1_l_1000000.cpu.pb.gz)
[m_1_l_1000000.mem.BEFORE.pb.gz](https://github.com/prometheus/statsd_exporter/files/1674500/m_1_l_1000000.mem.BEFORE.pb.gz)
[m_1_l_1000000.mem.pb.gz](https://github.com/prometheus/statsd_exporter/files/1674501/m_1_l_1000000.mem.pb.gz)
[m_1000000_l_1.cpu.BEFORE.pb.gz](https://github.com/prometheus/statsd_exporter/files/1674502/m_1000000_l_1.cpu.BEFORE.pb.gz)
[m_1000000_l_1.cpu.pb.gz](https://github.com/prometheus/statsd_exporter/files/1674503/m_1000000_l_1.cpu.pb.gz)
[m_1000000_l_1.mem.BEFORE.pb.gz](https://github.com/prometheus/statsd_exporter/files/1674504/m_1000000_l_1.mem.BEFORE.pb.gz)
[m_1000000_l_1.mem.pb.gz](https://github.com/prometheus/statsd_exporter/files/1674505/m_1000000_l_1.mem.pb.gz)

Bottom line: With the changes here, the profile looks broadly the same, whether you collect 1M collectors with one metric or one collector with 1M metrics.