@beorn7 2018-01-29
Just checked how many goroutines were started for gathering with this change:
To collect from one collector, indeed only one.
To collect from 1M collectors, it started 11 goroutines.

This seems to work out quite nicely.