@beorn7 2018-01-30
I added a safety measure to not schedule a large or even infinite number of goroutines (which can happen in theory).

With the benchmark results in, I think this works great and we should merge. @matthiasr WDYT?